fix: only publish release after all image builds succeed (#77)
While splitting the image publishing out into separate jobs for AWS, Azure and GCP, I noticed that we publish the new release as part of the AWS publish, not when all the image publishes have completed.

I've adjusted the workflows so that we do the release creation as a separate job that relies on all three images to successfully publish before creating the new release.